+++ ibas.materials/src/main/java/org/colorcoding/ibas/materials/logic/MaterialIssueService.java
@@ -3,17 +3,12 @@ package org.colorcoding.ibas.materials.logic;
 import java.math.BigDecimal;
 
 import org.colorcoding.ibas.bobas.approval.IApprovalData;
-import org.colorcoding.ibas.bobas.bo.IBODocument;
-import org.colorcoding.ibas.bobas.bo.IBODocumentLine;
-import org.colorcoding.ibas.bobas.bo.IBOTagCanceled;
-import org.colorcoding.ibas.bobas.bo.IBOTagDeleted;
 import org.colorcoding.ibas.bobas.common.ConditionOperation;
 import org.colorcoding.ibas.bobas.common.ConditionRelationship;
 import org.colorcoding.ibas.bobas.common.Criteria;
 import org.colorcoding.ibas.bobas.common.ICondition;
 import org.colorcoding.ibas.bobas.common.ICriteria;
 import org.colorcoding.ibas.bobas.common.IOperationResult;
-import org.colorcoding.ibas.bobas.core.ITrackStatus;
 import org.colorcoding.ibas.bobas.data.Decimal;
 import org.colorcoding.ibas.bobas.data.emApprovalStatus;
 import org.colorcoding.ibas.bobas.data.emDirection;
@@ -70,13 +65,8 @@ public class MaterialIssueService
 		}
 		boolean status = super.checkDataStatus(data);
 		if (status == false && this.isEnableMaterialCosts()) {
-			// 取消和标记删除时，执行逻辑
-			if (this.getRoot() == data || this.getHost() == data) {
-				status = super.checkDataStatus(data, ITrackStatus.class, IBOTagCanceled.class, IBOTagDeleted.class,
-						IBODocument.class, IBODocumentLine.class);
-			} else {
-				status = super.checkDataStatus(data, ITrackStatus.class, IBOTagCanceled.class, IBOTagDeleted.class);
-			}
+			// 激活成本计算时，正向逻辑都执行
+			status = true;
 		}
 		return status;
 	}
@@ -171,18 +161,20 @@ public class MaterialIssueService
 					throw new BusinessLogicException(operationResult.getError());
 				}
 				if (operationResult.getResultObjects().isEmpty()) {
-					throw new BusinessLogicException(I18N.prop("msg_mm_document_not_found_receipt_journal",
-							String.format("{[%s].[DocEntry = %s]%s}", contract.getDocumentType(),
-									contract.getDocumentEntry(),
-									contract.getDocumentLineId() > 0
-											? String.format("&&[LineId = %s]", contract.getDocumentLineId())
-											: "")));
+					materialJournal = new MaterialInventoryJournal();
+					((MaterialInventoryJournal) materialJournal).unsavable();
+					materialJournal.setBaseDocumentType(contract.getDocumentType());
+					materialJournal.setBaseDocumentEntry(contract.getDocumentEntry());
+					materialJournal.setBaseDocumentLineId(contract.getDocumentLineId());
+					materialJournal.setDirection(emDirection.OUT);
+					materialJournal.setDataSource(DATASOURCE_SIGN_OFFSETTING_JOURNAL);
+				} else {
+					materialJournal = operationResult.getResultObjects().firstOrDefault();
+					materialJournal = ((MaterialInventoryJournal) materialJournal).clone();
+					materialJournal.setDataSource(DATASOURCE_SIGN_OFFSETTING_JOURNAL);
+					materialJournal.setQuantity(materialJournal.getQuantity().negate());
+					materialJournal.setTransactionValue(materialJournal.getTransactionValue().negate());
 				}
-				materialJournal = operationResult.getResultObjects().firstOrDefault();
-				materialJournal = ((MaterialInventoryJournal) materialJournal).clone();
-				materialJournal.setDataSource(DATASOURCE_SIGN_OFFSETTING_JOURNAL);
-				materialJournal.setQuantity(materialJournal.getQuantity().negate());
-				materialJournal.setTransactionValue(materialJournal.getTransactionValue().negate());
 			} else {
 				materialJournal = new MaterialInventoryJournal();
 				materialJournal.setBaseDocumentType(contract.getDocumentType());
@@ -210,6 +202,10 @@ public class MaterialIssueService
 			}
 		}
 		IMaterialInventoryJournal materialJournal = this.getBeAffected();
+		// 不可保存对象，不执行逻辑
+		if (!materialJournal.isSavable()) {
+			return;
+		}
 		// 开启物料成本计算
 		if (this.isEnableMaterialCosts()) {
 			if (!materialJournal.isNew() || (contract.isOffsetting() && materialJournal.isNew())) {
@@ -379,6 +375,10 @@ public class MaterialIssueService
 	@Override
 	protected void revoke(IMaterialIssueContract contract) {
 		IMaterialInventoryJournal materialJournal = this.getBeAffected();
+		// 不可保存对象，不执行逻辑
+		if (!materialJournal.isSavable()) {
+			return;
+		}
 		if (!this.isEnableMaterialCosts() || contract.isOffsetting()) {
 			// 未开启成本的，删除
 			materialJournal.delete();
