+++ ibas.materials/src/main/java/org/colorcoding/ibas/materials/logic/MaterialSerialJournalService.java
@@ -3,17 +3,12 @@ package org.colorcoding.ibas.materials.logic;
 import java.math.BigDecimal;
 
 import org.colorcoding.ibas.bobas.approval.IApprovalData;
-import org.colorcoding.ibas.bobas.bo.IBODocument;
-import org.colorcoding.ibas.bobas.bo.IBODocumentLine;
-import org.colorcoding.ibas.bobas.bo.IBOTagCanceled;
-import org.colorcoding.ibas.bobas.bo.IBOTagDeleted;
 import org.colorcoding.ibas.bobas.common.ConditionOperation;
 import org.colorcoding.ibas.bobas.common.ConditionRelationship;
 import org.colorcoding.ibas.bobas.common.Criteria;
 import org.colorcoding.ibas.bobas.common.ICondition;
 import org.colorcoding.ibas.bobas.common.ICriteria;
 import org.colorcoding.ibas.bobas.common.IOperationResult;
-import org.colorcoding.ibas.bobas.core.ITrackStatus;
 import org.colorcoding.ibas.bobas.data.Decimal;
 import org.colorcoding.ibas.bobas.data.emApprovalStatus;
 import org.colorcoding.ibas.bobas.data.emDirection;
@@ -82,13 +77,8 @@ public class MaterialSerialJournalService
 		}
 		boolean status = super.checkDataStatus(data);
 		if (status == false && this.isEnableMaterialCosts()) {
-			// 取消和标记删除时，执行逻辑
-			if (this.getRoot() == data || this.getParent() == data) {
-				status = super.checkDataStatus(data, ITrackStatus.class, IBOTagCanceled.class, IBOTagDeleted.class,
-						IBODocument.class, IBODocumentLine.class);
-			} else {
-				status = super.checkDataStatus(data, ITrackStatus.class, IBOTagCanceled.class, IBOTagDeleted.class);
-			}
+			// 激活成本计算时，正向逻辑都执行
+			status = true;
 		}
 		return status;
 	}
@@ -175,18 +165,22 @@ public class MaterialSerialJournalService
 					throw new BusinessLogicException(operationResult.getError());
 				}
 				if (operationResult.getResultObjects().isEmpty()) {
-					throw new BusinessLogicException(I18N.prop("msg_mm_document_not_found_receipt_journal",
-							String.format("{[%s].[DocEntry = %s]%s}", contract.getDocumentType(),
-									contract.getDocumentEntry(),
-									contract.getDocumentLineId() > 0
-											? String.format("&&[LineId = %s]", contract.getDocumentLineId())
-											: "")));
+					materialSerialJournal = new MaterialSerialJournal();
+					((MaterialSerialJournal) materialSerialJournal).unsavable();
+					materialSerialJournal.setDirection(contract.getDirection());
+					materialSerialJournal.setBaseDocumentType(contract.getDocumentType());
+					materialSerialJournal.setBaseDocumentEntry(contract.getDocumentEntry());
+					materialSerialJournal.setBaseDocumentLineId(contract.getDocumentLineId());
+					materialSerialJournal.setBaseDocumentIndex(contract.getDocumentIndex());
+					materialSerialJournal.setDataSource(DATASOURCE_SIGN_OFFSETTING_JOURNAL);
+				} else {
+					materialSerialJournal = operationResult.getResultObjects().firstOrDefault();
+					materialSerialJournal = ((MaterialSerialJournal) materialSerialJournal).clone();
+					materialSerialJournal.setDataSource(DATASOURCE_SIGN_OFFSETTING_JOURNAL);
+					materialSerialJournal.setQuantity(materialSerialJournal.getQuantity().negate());
+					materialSerialJournal.setTransactionValue(materialSerialJournal.getTransactionValue().negate());
+
 				}
-				materialSerialJournal = operationResult.getResultObjects().firstOrDefault();
-				materialSerialJournal = ((MaterialSerialJournal) materialSerialJournal).clone();
-				materialSerialJournal.setDataSource(DATASOURCE_SIGN_OFFSETTING_JOURNAL);
-				materialSerialJournal.setQuantity(materialSerialJournal.getQuantity().negate());
-				materialSerialJournal.setTransactionValue(materialSerialJournal.getTransactionValue().negate());
 			} else {
 				materialSerialJournal = new MaterialSerialJournal();
 				materialSerialJournal.setDirection(contract.getDirection());
@@ -203,6 +197,10 @@ public class MaterialSerialJournalService
 	@Override
 	protected void impact(IMaterialSerialJournalContract contract) {
 		IMaterialSerialJournal materialSerialJournal = this.getBeAffected();
+		// 不可保存对象，不执行逻辑
+		if (!materialSerialJournal.isSavable()) {
+			return;
+		}
 		// 开启物料成本计算
 		if (this.isEnableMaterialCosts()) {
 			String localCurrency = org.colorcoding.ibas.accounting.MyConfiguration
@@ -417,6 +415,10 @@ public class MaterialSerialJournalService
 	@Override
 	protected void revoke(IMaterialSerialJournalContract contract) {
 		IMaterialSerialJournal materialSerialJournal = this.getBeAffected();
+		// 不可保存对象，不执行逻辑
+		if (!materialSerialJournal.isSavable()) {
+			return;
+		}
 		if (!this.isEnableMaterialCosts() || contract.isOffsetting()) {
 			// 未开启成本的，删除
 			materialSerialJournal.delete();
